Fixxed internal width error of PyQT and layouts with a workaround.
To be honest: There is safe a better way to do so. But for the moment it's good enough, that the width of a layout is not automatically > 600px when it's not fixxed manually. You can see that problem if you call self.width() in the class Clipper without these changes.